requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
openai==2.6.1
pillow==12.0.0
opencv-python==4.12.0.88
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.3